mathew: fix rename of `pvc-viewer` path
