+++ components/pvcviewer-controller/PROJECT
@@ -1,7 +1,7 @@
 domain: kubeflow.org
 layout:
 - go.kubebuilder.io/v4-alpha
-projectName: pvcviewer-controller
+projectName: pvc-viewer
 repo: github.com/kubeflow/notebooks/components/pvcviewer-controller
 resources:
 - api:
